// lib/earth3d/proto/rocktree.ts
export interface INodeKey {
    path?: string;
    epoch?: number;
}

export interface ICopyrightRequest {
    epoch?: number;
}

export interface ITextureDataRequest {
    node_key?: INodeKey;
    texture_format?: Texture_Format_Value;
    view_direction?: Texture_ViewDirection_Value;
}

export interface IBulkMetadata {
    node_metadata?: INodeMetadata[];
    head_node_key?: INodeKey;
    head_node_center?: number[];
    meters_per_texel?: number[];
    default_imagery_epoch?: number;
    default_available_texture_formats?: number;
    default_available_view_dependent_textures?: number;
    default_available_view_dependent_texture_formats?: number;
}

export interface INodeMetadata {
    path_and_flags?: number;
    epoch?: number;
    bulk_metadata_epoch?: number;
    oriented_bounding_box?: Uint8Array;
    meters_per_texel?: number;
    processing_oriented_bounding_box?: number[];
    imagery_epoch?: number;
    available_texture_formats?: number;
    available_view_dependent_textures?: number;
    available_view_dependent_texture_formats?: number;
}

export type NodeMetadata_Flags_Key = "RICH3D_LEAF" | "RICH3D_NODATA" | "LEAF" | "NODATA" | "USE_IMAGERY_EPOCH";
export type NodeMetadata_Flags_Value = 1 | 2 | 4 | 8 | 16;

export interface INodeData {
    matrix_globe_from_mesh?: number[];
    meshes?: IMesh[];
    copyright_ids?: number[];
    node_key?: INodeKey;
    kml_bounding_box?: number[];
    water_mesh?: IMesh;
    overlay_surface_meshes?: IMesh[];
    for_normals?: Uint8Array;
}

export interface IMesh {
    vertices?: Uint8Array;
    vertex_alphas?: Uint8Array;
    texture_coords?: Uint8Array;
    indices?: Uint8Array;
    octant_ranges?: Uint8Array;
    layer_counts?: Uint8Array;
    texture?: ITexture[];
    texture_coordinates?: Uint8Array;
    uv_offset_and_scale?: number[];
    layer_and_octant_counts?: Uint8Array;
    normals?: Uint8Array;
    normals_dev?: Uint8Array;
    mesh_id?: number;
    skirt_flags?: Uint8Array;
}
// ...
export interface IPlanetoidMetadata {
    root_node_metadata?: INodeMetadata;
    radius?: number;
    min_terrain_altitude?: number;
    max_terrain_altitude?: number;
}
// ...
export const NodeKey = {
    read(pbf: Pbf, end?: number): INodeKey {
        return pbf.readFields(NodeKey._readField, {path: "", epoch: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.path = pbf.readString();
        else if (tag === 2) obj.epoch = pbf.readVarint();
    },
    write(obj: INodeKey, pbf: Pbf): void {
        if (obj.path) pbf.writeStringField(1, obj.path);
        if (obj.epoch) pbf.writeVarintField(2, obj.epoch);
    }
};
// ...
export const BulkMetadata = {
    read(pbf: Pbf, end?: number): IBulkMetadata {
        return pbf.readFields(BulkMetadata._readField, {node_metadata: [], head_node_key: null, head_node_center: [], meters_per_texel: [], default_imagery_epoch: 0, default_available_texture_formats: 0, default_available_view_dependent_textures: 0, default_available_view_dependent_texture_formats: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.node_metadata.push(NodeMetadata.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 2) obj.head_node_key = NodeKey.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 3) pbf.readPackedDouble(obj.head_node_center);
        else if (tag === 4) pbf.readPackedFloat(obj.meters_per_texel);
        else if (tag === 5) obj.default_imagery_epoch = pbf.readVarint();
        else if (tag === 6) obj.default_available_texture_formats = pbf.readVarint();
        else if (tag === 7) obj.default_available_view_dependent_textures = pbf.readVarint();
        else if (tag === 8) obj.default_available_view_dependent_texture_formats = pbf.readVarint();
    },
    write(obj: IBulkMetadata, pbf: Pbf): void {
        if (obj.node_metadata) for (var i = 0; i < obj.node_metadata.length; i++) pbf.writeMessage(1, NodeMetadata.write, obj.node_metadata[i]);
        if (obj.head_node_key) pbf.writeMessage(2, NodeKey.write, obj.head_node_key);
        if (obj.head_node_center) pbf.writePackedDouble(3, obj.head_node_center);
        if (obj.meters_per_texel) pbf.writePackedFloat(4, obj.meters_per_texel);
        if (obj.default_imagery_epoch) pbf.writeVarintField(5, obj.default_imagery_epoch);
        if (obj.default_available_texture_formats) pbf.writeVarintField(6, obj.default_available_texture_formats);
        if (obj.default_available_view_dependent_textures) pbf.writeVarintField(7, obj.default_available_view_dependent_textures);
        if (obj.default_available_view_dependent_texture_formats) pbf.writeVarintField(8, obj.default_available_view_dependent_texture_formats);
    }
};

export const NodeMetadata = {

    read(pbf: Pbf, end?: number): INodeMetadata {
        return pbf.readFields(NodeMetadata._readField, {path_and_flags: 0, epoch: 0, bulk_metadata_epoch: 0, oriented_bounding_box: null, meters_per_texel: 0, processing_oriented_bounding_box: [], imagery_epoch: 0, available_texture_formats: 0, available_view_dependent_textures: 0, available_view_dependent_texture_formats: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.path_and_flags = pbf.readVarint();
        else if (tag === 2) obj.epoch = pbf.readVarint();
        else if (tag === 5) obj.bulk_metadata_epoch = pbf.readVarint();
        else if (tag === 3) obj.oriented_bounding_box = pbf.readBytes();
        else if (tag === 4) obj.meters_per_texel = pbf.readFloat();
        else if (tag === 6) pbf.readPackedDouble(obj.processing_oriented_bounding_box);
        else if (tag === 7) obj.imagery_epoch = pbf.readVarint();
        else if (tag === 8) obj.available_texture_formats = pbf.readVarint();
        else if (tag === 9) obj.available_view_dependent_textures = pbf.readVarint();
        else if (tag === 10) obj.available_view_dependent_texture_formats = pbf.readVarint();
    },
    write(obj: INodeMetadata, pbf: Pbf): void {
        if (obj.path_and_flags) pbf.writeVarintField(1, obj.path_and_flags);
        if (obj.epoch) pbf.writeVarintField(2, obj.epoch);
        if (obj.bulk_metadata_epoch) pbf.writeVarintField(5, obj.bulk_metadata_epoch);
        if (obj.oriented_bounding_box) pbf.writeBytesField(3, obj.oriented_bounding_box);
        if (obj.meters_per_texel) pbf.writeFloatField(4, obj.meters_per_texel);
        if (obj.processing_oriented_bounding_box) pbf.writePackedDouble(6, obj.processing_oriented_bounding_box);
        if (obj.imagery_epoch) pbf.writeVarintField(7, obj.imagery_epoch);
        if (obj.available_texture_formats) pbf.writeVarintField(8, obj.available_texture_formats);
        if (obj.available_view_dependent_textures) pbf.writeVarintField(9, obj.available_view_dependent_textures);
        if (obj.available_view_dependent_texture_formats) pbf.writeVarintField(10, obj.available_view_dependent_texture_formats);
    },

    Flags: {
    "RICH3D_LEAF": {
        "value": 1,
        "options": {}
    },
    "RICH3D_NODATA": {
        "value": 2,
        "options": {}
    },
    "LEAF": {
        "value": 4,
        "options": {}
    },
    "NODATA": {
        "value": 8,
        "options": {}
    },
    "USE_IMAGERY_EPOCH": {
        "value": 16,
        "options": {}
    }
} as { [K in NodeMetadata_Flags_Key]: { value: NodeMetadata_Flags_Value, options: any } }


};

export const NodeData = {
    read(pbf: Pbf, end?: number): INodeData {
        return pbf.readFields(NodeData._readField, {matrix_globe_from_mesh: [], meshes: [], copyright_ids: [], node_key: null, kml_bounding_box: [], water_mesh: null, overlay_surface_meshes: [], for_normals: null}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) pbf.readPackedDouble(obj.matrix_globe_from_mesh);
        else if (tag === 2) obj.meshes.push(Mesh.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 3) pbf.readPackedVarint(obj.copyright_ids);
        else if (tag === 4) obj.node_key = NodeKey.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 5) pbf.readPackedDouble(obj.kml_bounding_box);
        else if (tag === 6) obj.water_mesh = Mesh.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 7) obj.overlay_surface_meshes.push(Mesh.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 8) obj.for_normals = pbf.readBytes();
    },
    write(obj: INodeData, pbf: Pbf): void {
        if (obj.matrix_globe_from_mesh) pbf.writePackedDouble(1, obj.matrix_globe_from_mesh);
        if (obj.meshes) for (var i = 0; i < obj.meshes.length; i++) pbf.writeMessage(2, Mesh.write, obj.meshes[i]);
        if (obj.copyright_ids) for (i = 0; i < obj.copyright_ids.length; i++) pbf.writeVarintField(3, obj.copyright_ids[i]);
        if (obj.node_key) pbf.writeMessage(4, NodeKey.write, obj.node_key);
        if (obj.kml_bounding_box) pbf.writePackedDouble(5, obj.kml_bounding_box);
        if (obj.water_mesh) pbf.writeMessage(6, Mesh.write, obj.water_mesh);
        if (obj.overlay_surface_meshes) for (i = 0; i < obj.overlay_surface_meshes.length; i++) pbf.writeMessage(7, Mesh.write, obj.overlay_surface_meshes[i]);
        if (obj.for_normals) pbf.writeBytesField(8, obj.for_normals);
    }
};
// ...
export const PlanetoidMetadata = {
    read(pbf: Pbf, end?: number): IPlanetoidMetadata {
        return pbf.readFields(PlanetoidMetadata._readField, {root_node_metadata: null, radius: 0, min_terrain_altitude: 0, max_terrain_altitude: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.root_node_metadata = NodeMetadata.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 2) obj.radius = pbf.readFloat();
        else if (tag === 3) obj.min_terrain_altitude = pbf.readFloat();
        else if (tag === 4) obj.max_terrain_altitude = pbf.readFloat();
    },
    write(obj: IPlanetoidMetadata, pbf: Pbf): void {
        if (obj.root_node_metadata) pbf.writeMessage(1, NodeMetadata.write, obj.root_node_metadata);
        if (obj.radius) pbf.writeFloatField(2, obj.radius);
        if (obj.min_terrain_altitude) pbf.writeFloatField(3, obj.min_terrain_altitude);
        if (obj.max_terrain_altitude) pbf.writeFloatField(4, obj.max_terrain_altitude);
    }
};